+++ changelog.md
@@ -1,4 +1,4 @@
-## JO'N Nov 20 2020
+conv## JO'N Nov 20 2020
 - Fixed a bug in ConvertTo-PowerShellNotebook (missing -raw)
 - Fixed a bug in ConvertToPowerShellNotebookTests (test-pat with no H)
 - Fixed a couple of other test issues, including skipping SQL tests if SQL server module is not present.
@@ -7,7 +7,7 @@
 - Added Language and presence of parameterized cells to Get-Notebook results. 
 - Created a new ConvertFrom-IPYNB which incorporates the convertFrom-NotebookToMarkdown, and export-notebookToPowerShell script functionality and adds convert to html as well.
 - In several places I've changed the name of "path to notebook" parameters to be `path` (with the old name and "fullname" as aliases) and added ValueFromPipelineByProperty name so items can be pipe items in, and also ensured path can be a list and contain wildcards.
-- Added -Recurse and support for URLs as paths to Get-Notebook. 
+- Added -Recurse to Get-Notebook. 
 
 ## JO'N Oct  1 2020
 - Export NotebookToPowerShellScript